return logic in screen
