--- src/components/singin-store-wallet/index.tsx
@@ -1,1 +1,0 @@
-export * from './singin-store-wallet';
